% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_df_input}
\alias{check_df_input}
\title{Check dataframe input}
\usage{
check_df_input(df)
}
\arguments{
\item{df}{Dataframe returned from get_neighbour_zestimates()}
}
\description{
Checks that input is a dataframe, has at least 1 column,
has column names "latitude", "longitude", "zestimate", "address"
}
\keyword{internal}
