% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_response_type}
\alias{check_response_type}
\title{Exception handler: response type}
\usage{
check_response_type(response)
}
\arguments{
\item{response}{API response}
}
\description{
Check if the response is in XML format.
}
\keyword{internal}
