% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Exception handler: response status}
\usage{
check_status(response)
}
\arguments{
\item{response}{API response}
}
\description{
Check if the API request is successful. Note that Zillow API always return `200`,
and the exact request status and message is buried inside the response.
}
\details{
Throw errors if request is not successful and return the failed request code and messages.
}
\keyword{internal}
