% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_params.R
\name{format_address}
\alias{format_address}
\title{Format Address}
\usage{
format_address(address)
}
\arguments{
\item{address}{Address as a string (e.g., '2144 Bigelow Ave')}
}
\value{
proper format of address as a string (e.g., '2144+Bigelow+Ave')
}
\description{
Converts the property address to the query param format
}
\keyword{internal}
