% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_params.R
\name{format_citystate}
\alias{format_citystate}
\title{Format City and State}
\usage{
format_citystate(city, state)
}
\arguments{
\item{city}{City of property (e.g., 'Seattle')}

\item{state}{State of property. Can be abbreviated (e.g., 'WA' for Washington)}
}
\value{
proper format of city/state as a string (e.g., 'Seattle%2C+WA')
}
\description{
Format city and state query parameter
}
\keyword{internal}
