% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{get_comp_df}
\alias{get_comp_df}
\title{Get comparison data frame}
\usage{
get_comp_df(zpid, count = 25)
}
\arguments{
\item{zpid}{property ID}

\item{count}{number of comparable properties (max 25)}
}
\value{
A data frame with n comparable properties and some of their main attributes.
}
\description{
Retrieves some of the attributes of comparable properties estimated by Zillow.
}
