% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{get_deep_comps}
\alias{get_deep_comps}
\title{GetDeepComps Response}
\usage{
get_deep_comps(zpid, count)
}
\arguments{
\item{zpid}{property ID}

\item{count}{number of comparable properties}
}
\value{
A data frame with n comparable properties and some of their main attributes.
}
\description{
Retrieves GetDeepComps API response.
}
