% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lat}
\alias{get_lat}
\title{Get latitude of property}
\usage{
get_lat(response)
}
\arguments{
\item{response}{Object returned from get_search_results()}
}
\value{
Latitude as a float
}
\description{
Gets latitude coordinate of a given property
}
\keyword{internal}
