% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_long}
\alias{get_long}
\title{Get longitude of property}
\usage{
get_long(response)
}
\arguments{
\item{response}{Object returned from get_search_results()}
}
\value{
Longitude as a float
}
\description{
Gets longitude coordinate of a given property
}
\keyword{internal}
