% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_near}
\alias{get_near}
\title{Get local real estate data from the API response}
\usage{
get_near(response)
}
\arguments{
\item{response}{The API response from `get_search_results`}
}
\value{
A data frame with local real estate data
\itemize{
  \item Region of the property
  \item Region id
  \item Region type
  \item Zillow Home Value Index
}
}
\description{
Get local real estate data, e.g. Region, id, type(i.e. neighborhood), Zillow Home Value Index
}
\examples{
\dontrun{
response <- get_search_results("2144 Bigelow Ave", "Seattle", "WA")
get_near(response)

get_search_results("2144 Bigelow Ave", "Seattle", "WA") \%>\%
  get_near()
}
}
