% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_neighbour_zestimates}
\alias{get_neighbour_zestimates}
\title{Get Neighbour Zestimates}
\usage{
get_neighbour_zestimates(address, city, state)
}
\arguments{
\item{address}{Street address of interest as a string
(e.g., '2144 Bigelow Ave')}

\item{city}{City of the property}

\item{state}{State of the property. Can be abbreviated (e.g., 'WA' for Washington)}
}
\value{
Dataframe with neighbour address, latitude and longitude coordinates,
and corresponding zestimate
}
\description{
Gets zestimates of neighbouring houses on the same street.
}
