% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_search_results}
\alias{get_search_results}
\title{GetSearchResults Response}
\usage{
get_search_results(address, citystatezip)
}
\arguments{
\item{address}{Street address of interest as a string (e.g., '2144+Bigelow+Ave')}

\item{citystatezip}{either the city and state or ZIP code of the street address (e.g., 'Seattle%2C+WA')}
}
\value{
API response
}
\description{
Retrieves GetSearchResults API response.
}
