% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_zestimate}
\alias{get_zestimate}
\title{Get Zestimate}
\usage{
get_zestimate(zwsid, zpid)
}
\arguments{
\item{zwsid}{Access token}

\item{zpid}{Zillow property ID (zpid)}
}
\value{
Zestimate of house (in USD)
}
\description{
Retrives zestimate information for a specified zpid.
}
