% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_zestimate_all}
\alias{get_zestimate_all}
\title{Get Zestimate related data from the API response}
\usage{
get_zestimate_all(response)
}
\arguments{
\item{response}{The API response from `get_search_results`}
}
\value{
A data frame including the following information:
\itemize{
  \item Currency
  \item Zestimate
  \item Last updated
  \item 30-day change
  \item Valuation range (low & high)
}
}
\description{
The Zestimate® home value is Zillow's estimated market value for
an individual home and is calculated for about 100 million homes nationwide.
It is a starting point in determining a home's value and is not an official appraisal.
The Zestimate is automatically computed daily based on millions of public and
user-submitted data points.
}
\examples{
\dontrun{
response <- get_search_results("2144 Bigelow Ave", "Seattle", "WA")
get_zestimate_all(response)

get_search_results("2144 Bigelow Ave", "Seattle", "WA") \%>\%
  get_zestimate_all()
}
}
