% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_zpid}
\alias{get_zpid}
\title{Get Zillow Property ID}
\usage{
get_zpid(address, city, state)
}
\arguments{
\item{address}{Street address of interest as a string
(e.g., '2144+Bigelow+Ave')}

\item{city}{City of the property}

\item{state}{State of the property. Can be abbreviated (e.g., 'WA' for Washington)}
}
\value{
zpid as a string
}
\description{
Retrieves the Zillow Property ID (zpid) for a house
from the GetSearchResults API.
}
