% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{plot_neighbour_zestimates}
\alias{plot_neighbour_zestimates}
\title{Plot neighbour zestimates}
\usage{
plot_neighbour_zestimates(df)
}
\arguments{
\item{df}{Dataframe returned from get_neighbour_zestimates()}
}
\value{
A ggplot map
}
\description{
Plot neighbour zestimates
}
