% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{price_plot}
\alias{price_plot}
\title{Price ranges plot}
\usage{
price_plot(zpid)
}
\arguments{
\item{zpid}{property ID}
}
\value{
A ggplot boxplot with the price ranges of similar properties.
}
\description{
Creates a boxplot with the price ranges of similar properties.
}
