% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{price_ranges}
\alias{price_ranges}
\title{Price and rent ranges.}
\usage{
price_ranges(zpid)
}
\arguments{
\item{zpid}{property ID}
}
\value{
A list with two elements: similar prices and similar rents.
}
\description{
Retrieves some estimates for similar rent and house prices for a certain property.
}
