% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realestateR.R
\docType{package}
\name{RealEstateR}
\alias{RealEstateR}
\alias{RealEstateR-package}
\title{\code{RealEstateR} package}
\description{
RealEstateR: a Zillow API Wrapper
}
\details{
See the README on
\href{https://github.com/estebanangelm/RealEstateR#readme}{GitHub}
}
