% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{reviews}
\alias{reviews}
\title{Get Review Details of Agents of your Choice}
\usage{
reviews(screennames)
}
\arguments{
\item{screennames}{Agent screenname(s) of your choice (at most 5 screennames)}
}
\value{
A dataframe that includes details of agents, and their star reviews.
}
\description{
This function inputs up to 5 agents' screenname,
then output the agents' information and their review stars.
}
\examples{
\dontrun{screennames <- c("mwalley0", "pamelarporter", "klamping4", "Cincysrealtor")
reviews(screennames)}

}
\author{
Ha Dinh
}
