% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews_get_screenname.R
\name{reviews_get_screennames}
\alias{reviews_get_screennames}
\title{Get Agent IDs Based On Location}
\usage{
reviews_get_screennames(city, state)
}
\arguments{
\item{city}{A city of interest as a string.
(e.g. "Cincinnati", "Los Angeles")}

\item{state}{A state abbreviation of interest, where the city is in, as a string in 2 letters form.
(e.g. "OH")}
}
\value{
A dataframe that includes name, phone number, and id of agents in the location input.
}
\description{
This function inputs a combination of city and state of your choice,
then output a dataframe that includes name, screenname, phone number,
city and state of agents from the input location.
}
\examples{
\dontrun{df <- reviews_get_screennames('Los Angeles', 'CA')}

}
\author{
Ha Dinh
}
