% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_search_results}
\alias{get_search_results}
\title{GetSearchResults Response}
\usage{
get_search_results(address, city, state)
}
\arguments{
\item{address}{Street address of interest as a string
(e.g., '2144 Bigelow Ave')}

\item{city}{City of the property}

\item{state}{State of the property. Can be abbreviated (e.g., 'WA' for Washington)}
}
\value{
API response
}
\description{
Retrieves GetSearchResults API response.
}
