% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_zestimate}
\alias{get_zestimate}
\title{Get Zestimate}
\usage{
get_zestimate(zpid)
}
\arguments{
\item{zpid}{property ID}
}
\value{
Estimated price of house as a numeric (in USD)
}
\description{
Retrieves zestimate information for a specified zpid.
}
