% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_zpid}
\alias{get_zpid}
\title{Get Zillow Property ID}
\usage{
get_zpid(zwsid, address, citystatezip)
}
\arguments{
\item{zwsid}{Access token as a string}

\item{address}{Street address of interest as a string
(e.g., '2144+Bigelow+Ave')}

\item{citystatezip}{either the city and state or ZIP code
of the street address (e.g., 'Seattle%2C+WA')}
}
\value{
zpid as a string
}
\description{
Retrieves the Zillow Property ID (zpid) for a house
from the GetSearchResults API.
}
