% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zestimate.R
\name{get_zpid}
\alias{get_zpid}
\title{Get Zillow Property ID}
\usage{
get_zpid(response)
}
\arguments{
\item{response}{API response of get_search_results()}
}
\value{
zpid as a string
}
\description{
Retrieves the Zillow Property ID (zpid) for a house
from the GetSearchResults API.
}
\examples{
\dontrun{
response <- get_search_results("2144 Bigelow Ave", "Seattle", "WA")
get_zpid(response)
}

}
