% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{reviews}
\alias{reviews}
\title{Get Review Details of Agents in a Location}
\usage{
reviews(city, state)
}
\arguments{
\item{city}{A city of interest as a string.
(e.g. "Cincinnati" or "Los-Angeles")}

\item{state}{A state of interest, where the city is in, as a string in 2 letters form.
(e.g. "OH")}
}
\value{
A dataframe that includes details of agents, and their star reviews.
}
\description{
Get Review Details of Agents in a Location
}
\examples{
reviews("Cincinnati", "OH")

}
