% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews_get_screenname.R
\name{reviews_get_screennames}
\alias{reviews_get_screennames}
\title{Get Agent IDs Based On Location}
\usage{
reviews_get_screennames(name, city, state)
}
\arguments{
\item{name}{Full name of an agent/real-estate company you want to search for.}

\item{city}{A city of interest as a string.
(e.g. "Cincinnati", "Los Angeles")}

\item{state}{A state abbreviation of interest, where the city is in, as a string in 2 letters form.
(e.g. "OH")}
}
\value{
A string.
}
\description{
This function inputs a combination of agent name, city and state the agent is in,
then output the agent's screenname.
}
\examples{
\dontrun{reviews_get_screennames('Rakesh Ram Real Estate Group', 'Cincinnati', 'OH')}

}
\author{
Ha Dinh
}
